bugfix: CPEx10 SafeLoader support v1.1
Add safeloader support for CPEx10 v1.1 devices that are not UN market.

It appears there exists a CA,EU,UN and US models.

Still need to come back and add 2.0 devices but that will be done
as a feature add.

Jumpd the vendor field as precautionary related to 1.1 devices
to match existing vendor field in 2.0 firmware.

ref AREDN->ticket:183

Change-Id: I90ed649a819617321b4030165fddf2e990fbd53e